Change players to a linked list
Fix write_data() in ocr.c.
Improve cmd_source functions.

Add U32CAST to iterate over the Player struct for uint32_t fields.
Add a mutex for player_head.
Add find_player() to util.c.

Remove MAX_PLAYERS from config.def.h as it's not longer necessary.
Remove libre from the repo.
Remove xfopen() from util.c as it is in libre/x.h.

Change debug flags to main Makefile.
Change qsort to merge sort in cmd_leaderboard.c.
Change multiple boolean value from int to bool type.
Change Player struct to use uint32_t instead of long and to have name
and kingdom in the struct not as pointers.
Change MAX_USERNAME_SIZ and MAX_KINGDOM_SIZ to 32 for no struct padding.
Change try from libre/evil.h to the equivalent from libre/x.h.
Change twice in raids.c for time_run to try more than two times to
recognize the screenshot.
Move discord_send_message() to util.c.

+++ src/roles.c
@@ -59,7 +59,7 @@ level(struct discord *client, u64snowflake userid, Player *player)
 void
 playtime(struct discord *client, u64snowflake userid, Player *player)
 {
-	long playtime = player->playtime / 24;
+	uint32_t playtime = player->playtime / 24;
 
 	if (playtime >= 150) {
 		discord_add_guild_member_role(client, ROLE_GUILD_ID, userid,
